Modificação
criação de shop.

+++ P.I dos Cria/Assets/Script/EnemyHealth.cs
@@ -15,9 +15,9 @@ public class EnemyHealth : MonoBehaviour
         EnemyManager.Instance?.RegisterEnemy(gameObject);
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
-        currentHealth -= damage;
+        currentHealth -= (int)damage;
 
         if (currentHealth <= 0)
         {
